new chatbot and auto hear

+++ project_ws/src/cogrob_pepper_nodes/src/pepper_nodes/src/led_server.py
@@ -18,7 +18,7 @@ class LedNode:
         self.port = port
         self.session = Session(ip, port)
         self.lof = self.session.get_service("ALLeds")
-        self.DURATION=float(2) # <-- set duration
+        self.DURATION=float(1) # <-- set duration
